controladores v1.1.0
-codigo limpiado y reestructurado en algunos casos para mantener cierta consistencia.
-textos de message o parecidos en return eliminados, ahora devuelve un boolean de "success" para indicar si se ha realizado correctamente o no junto con un status.

+++ TalenTry/app/Http/Controllers/ControllerAnswer.php
@@ -14,26 +14,17 @@ class ControllerAnswer extends Controller
     {
         try {
             $answers = ModelAnswer::orderBy('questionID')->get();
+            if (!$answers) {
+                return response()->json(['success' => false], 404);
+            }
             return response()->json([$answers], 200);
         } catch (\Exception $e) {
-            return response()->json(['message' => 'Ha ocurrido un error al cargar las respuestas: ' . $e], 500);
+            return response()->json(['success' => false, 'error' => $e->getMessage()], 500);
         }
     }
-
-    //obtiene las respuestas a una pregunta, solo para funciones internas, no api!
-    public function showQuestion($id)
-    {
-        try {
-            $respuestas = ModelAnswer::where('QuestionID', $id)->get();
-            return $respuestas;
-        } catch (\Exception $e) {
-            return "Error";
-        }
-    }
-
 
     //crea una respuesta para la pregunta seleccionada
-    public function create()
+    public function create(Request $request)
     {
         try {
             $request->validate([
@@ -47,19 +38,19 @@ class ControllerAnswer extends Controller
                 'QuestionID' => $request->input('QuestionID'),
                 'QuestionPoints' => $request->input('QuestionPoints'),
             ]);
-                //save in database
+            //save in database
             if ($answer->save()) {
                 DB::commit();
-                return response()->json(['saved' => true], 200);
+                return response()->json(['success' => true], 200);
             }
         } catch (\Exception $e) {
             DB::rollBack();
-            return response()->json(['message' => 'Ha ocurrido un error al guardar la respuesta: ' . $e], 500);
+            return response()->json(['success' => false, 'error' => $e->getMessage()], 500);
         }
     }
 
     //actualiza la respuesta seleccionada
-    public function update($id)
+    public function update(Request $request, $id)
     {
         try {
             DB::beginTransaction();
@@ -71,7 +62,9 @@ class ControllerAnswer extends Controller
 
             // Validation logic goes here
             $answer = ModelAnswer::findOrFail($id);
-
+            if (!$answer) {
+                return response()->json(['success' => false], 404);
+            }
             // Use the update method to update the answer
             $answer->update([
                 'answer' => $request->input('answer'),
@@ -79,10 +72,10 @@ class ControllerAnswer extends Controller
                 'QuestionPoints' => $request->input('QuestionPoints'),
             ]);
             DB::commit();
-            return response()->json(['updated' => true], 200);
+            return response()->json(['success' => true], 200);
         } catch (\Exception $e) {
             DB::rollBack();
-            return response()->json(['message' => 'Ha ocurrido un error al actualizar la respuesta: ' . $e], 500);
+            return response()->json(['success' => false, 'error' => $e->getMessage()], 500);
         }
     }
 
@@ -91,17 +84,32 @@ class ControllerAnswer extends Controller
     {
         try {
             DB::beginTransaction();
-            
+
             $answer = ModelAnswer::findOrFail($id);
+            if (!$answer) {
+                return response()->json(['success' => false], 404);
+            }
             $answer->delete();
-            
+
             DB::commit();
-            
-            return response()->json(['success' => true, 'message' => 'Respuesta eliminada correctamente'],200);
+
+            return response()->json(['success' => true], 200);
         } catch (\Exception $e) {
             DB::rollBack();
-            
-            return response()->json(['success' => false, 'message' => 'No se pudo eliminar. '.$e],500);
+
+            return response()->json(['success' => false, 'error' => $e->getMessage()], 500);
         }
-    }  
+    }
+
+    //------------------------------------------------------------------------------
+    //obtiene las respuestas a una pregunta, solo para funciones internas, no api!!!
+    public function showQuestion($id)
+    {
+        try {
+            $respuestas = ModelAnswer::where('QuestionID', $id)->get();
+            return $respuestas;
+        } catch (\Exception $e) {
+            return "Error";
+        }
+    }
 }
